feat: Created utilities to support routine methods
