+++ src/App.css
@@ -9,10 +9,6 @@
     font-family: "Kanit", sans-serif;
 }
 
-/* #main-container__title {
-  z-index: 2;
-} */
-
 #main-container__title,
 #main-container__text {
     position: absolute;
