+++ src/utilities/launchWords.js
@@ -1,0 +1,17 @@
+// Variables
+const recipes = ["pbj", "chicken and broccoli stir-fry", "meatball sub"];
+
+// Command Words
+const greetings = [
+    "hello cleo",
+    "hello clio",
+    "hi cleo",
+    "hi clio",
+    "he cleo",
+    "he clio",
+    "hey cleo",
+    "hey clio",
+];
+const aborts = ["abort cleo", "abort clio"];
+
+export { recipes, greetings, aborts };
